+++ dds/src/implementation/mod.rs
@@ -3,3 +3,4 @@ pub mod actors;
 pub mod data_representation_builtin_endpoints;
 pub mod data_representation_inline_qos;
 pub mod runtime;
+pub mod xtypes_glue;
