+++ dds/src/xtypes/mod.rs
@@ -1,17 +1,10 @@
-pub mod deserialize;
-
-pub mod deserializer;
-
-pub mod serialize;
-
-pub mod serializer;
-
-pub mod dynamic_type;
-
-pub mod error;
-
 pub mod bytes;
-pub mod instance_handle;
+pub mod deserialize;
+pub mod deserializer;
+pub mod dynamic_type;
+pub mod error;
+pub mod serialize;
+pub mod serializer;
 pub mod type_object;
 pub mod xcdr_deserializer;
 pub mod xcdr_serializer;
